functions/transformation/requirements.txt
google-cloud-bigquery
google-cloud-storage
pandas